--- client/src/components/Landing/Circle/circle.css
@@ -1,13 +1,0 @@
-
-.circle {
-    border-radius: 50%;
-    width: 200px;
-    height: 200px;
-    border: 2px solid white;
-    margin: 20px 0px;
-}
-
-
-.circle-text {
-    color: white;
-}
