--- client/src/components/Landing/Steps/Steps.css
@@ -1,9 +1,0 @@
-.stepNum{
-    float: left;   
-    display: block;
-    color: #0d47a1;
-}
-
-.stepName {
-    margin: 40px 0px 0px 50px;
-}
